Add DB index for server side completion
This adds an extra table to the table to the db index.
This allows us to quickly query for server side completion data
without having to load botocore.  As a result, we can now defer
loading anything related to botocore until we know that we can
complete a value via server side completion.