Add top level file parameter for translate
The --import-terminology parameter takes a dictionary as a value, one of
the required keys in this dictionary Is File which is a binary blob
loaded from a file. This has been hoisted to the top level as the --file
parameter which will can be used with the fileb:// prefix to load a
bytes object and populate the dictionary.