Unroll YAML anchors in cloudformation package
This updates the YAML dumper for the package command to ignore aliases
/ anchors. For context, YAML lets you do something like:

```yaml
Foo: &anchor
  Bar: Baz
Bin:
  <<: *anchor
  Spam: Eggs
```

Which will result in the `Bin` key being a copy of the `Foo` key, but
with an additional member `Spam`. CloudFormation won't accept
templates that contain these, but they can be useful in the cases
where CloudFormation conditions aren't supported, causing you to have
to duplicate a resource with one or two changes.

With this change users can write their template using anchors and then
call package to resolve them along with anything else package needs to
do.