No longer special case Python 3.7
Now that the Xenial builds of Travis CI are officially supported, it makes sense to run all builds on it. Python 3.7 is not supported on Trusty.

Sudo is no longer required to run builds on Xenial, so we can set this globally back to `false` and not have to worry.