Get functional tests passing on windows
The streaming tests added a .files property to the base Params test
class that is used by many other test cases. The history recorder tests
in particular needed control of how the files were then cleaned up to
prevent file connections from preventing their deletion on
windows. Since the cleanup was added to the base class as well this was
now happening in the superclass before all the connections to the file
had been correctly closed in the subclass's tearDown method.

The fix was just to move the files property down to the specific test
subclass that needed it. All the subclasses of this do not need it and
we probably shouldn't modify a heavily used base class for functionality
needed by a very small set of subclasses.