Update EMR Configurations schema to support three-layer nested configurations
This change allows configurations given to instance groups and instance fleets
when creating an EMR cluster to be an up-to-three-layer nested structure with
the most inner layer of configuration always being empty. In this way, syntax
check of create-cluster command is consistent with current EMR API validation
for configuration objects.