Fixed eks get-token customization
When python base64 encodes bytes, it will add padding characters (equal
signs, `=`) to the end of strings that where the encoded length is not
divisible by 4. aws-iam-authenticator [1] uses go's implementation of
base64 url encoding [2] which strips padding characters. The difference
of behavior causes padded tokens to be rejected by the authenticator
server.

[1] https://github.com/kubernetes-sigs/aws-iam-authenticator
[2] https://golang.org/pkg/encoding/base64/#RawURLEncoding