Fix #3618 s3uploader writing to stdout
The package command writes the template out to stdout.  Having the
status written on stderr results in the standard practice of piping
stdout to a template file to break.

eg.  aws cloudformation package --template-file cloudformation.yaml
--s3-bucket goehd-test > cloudformation.pkg

This results in the value

\rUploading ...

In the first line of the output with deploy complaining about
mapped values not being allowed in <unicode string>.

Redirecting this to stderr reverts the behavior to be compliant with
standard practice.