Add validation checks to example files
This adds a `test_examples.py` script that verifies all of our
.rst files in `./examples`.  It's pretty basic right now but
it's already caught a number of issues.  It checks thst:

* Doc examples only have ascii chars
* Doc examples are all valid RST
* Doc examples reference commands that exist
* (Implicitly) Shell quoting/escaping is valid

The last command is fairly basic right now, it just checks that
the service and operation name map to something in a command table,
but we could also exand this to verify command line arguments
as well.

I've also renamed one of the files from an `.rst` to a `.txt`
extension.  These was used for the synopsis of a command, which
is formatted under a code block so it doesn't need to be (and isn't
in this case) valid rst.