Use new keys when testing metadata directive options
There were several cases tested for metadata directives.  This involved
copying an object with specific md values and checking what happens to the
various metadata on the copied object.  We were using the same
destination object for each of those cases.  This could occasionally
cause eventually consistency issues and test failures.  I've updated the
test to use a new destination object each time.  This is slightly
different behavior than what we were testing previously (checking if we
can wipe out existing metadata of an object vs. checking that metadata
is not copied to a new object), but given the test is about metadata
directives, I think from our perspective testing the behavior when
copying to a new object each time is still sufficient to show that we've
implemented this behavior correctly.