Added Windows OS support using PowerShell
The alias file syntax is designed for the
bash shell on linux. Running as-is causes
the following error:

$ aws myip
'f' is not recognized as an internal or external command,
operable program or batch file.

Using PowerShell with a few character
replacements allows the aliases to work.
The bash shell was considered but it needed
commands to end with a semicolon to work reliably
and it created an external dependency.