Add Mac PKG installer uninstall and test script
Uninstalling a Mac PKG is not straightforward, so a utility script has
been added that can uninstall the AWS CLI. The installer tester script
has been updated with a pkg format option. This requires sudo access for
now to both install and uninstall the package.

Another tangential change to the postinstall script was made to keep
a manifest of extra files created by the installation process. This
file is checked at uninstall time and used to clean up any files not
registered with the PKG receipt (the receipt cannot be updated manually
by a postinstall script, so this must be a separate metadata file).
Any new files added in the future should be added to this manifest as
well.