Remove tests for start param in route53domains
History of this change:

* The original args are "start" and "end"
* We couldn't use "end" because of the "--endpoint-url" shadowing.
* We renamed these args to "--start-time" and "--end-time"

What broke is that last year we added pagination for this operation,
which adds a "--starting-token", which means that this arg now shadows
the "--start" arg.

Given this is v2, I think it's cleaner to just require that this value
must either be "--start-time" and "--end-time" and remove support for
"--start".