Add ClusterArn to output of EMR create-cluster and add-instance-groups command.
Why is this change required?
This change will a new field, ClusterArn, to the response of the EMR create-cluster and add-instance-groups command.

How does this change solve the issue?  This change will update the create-cluster, add-instance-groups response to contain the new ClusterArn field.

Does this commit cause any expected behaviours to change? If so, what? No because this field is optional and nullable.