Fix timestamp related s3 tests failing on MacOS
On MacOS the file timestamp can no longer be invalid. This means our
tests that were checking for a warning by setting an invalid timestamp
no longer emit such a warning on newer Macs.

In one test it was specifically for invalid timestamps. This one is now
skipped on MacOS since it is not a valid case to check for anymore.

The other was a test to see if an error and a warning can be emitted at
the same time. It was using an invalid timestamp incidentally to emit a
warning. This test is run the same way on windows, and now skipped on
posix systems. The test was duplicated for posix systems and uses an
unreadable file to generate the warning.