Fix aws_completer to handle exe aliases
The completer was making the assumption that the command name
will always be aws. However, for developer preview the command
is aws2, which causes the completer to lookup aws2 in the
autocomplete index instead of aws. Now, the autocompleter
will not take into account the executable name when producing
completion results.

In addition, this fixes another issue where the completer did
not handle empty strings, which would cause index errors.