Refresh bundled installer to work with modern pip/venv
This updates the bundled installer generation to use modern versions of
virtualenv, and as a result, modern versions of pip, setuptools, etc.

Some of these changes may seem unnecessary, but here's why I had
to make all these changes.

The motivation for this change was that we couldn't upgrade our version
of dateutils in botocore because it used new functionality in its
build/install process that required a modern version of setuptools.  The
version we pull in comes from our virtualenv which is a fairly old
version.  So the first thing we had to do was upgrade our version of
virtualenv we use in the bundled installer.

That's where the fun begins.

First there's py26 issues, specifically that newer virtualenvs (>=16)
don't support python2.6.  So we have to branch and decide which version
of virtualenv to use based on our python version.  That's not really a
problem because new versions of dateutils don't support py26 anyways so
it doesn't change anything for py26.  We just need to add more branching
to the bundled installer.

Next we have the fact that because we're pulling in a new virtualenv
for > py26, we pull in a new pip.  A new, major version of pip, that is
not backwards compatible with the version we were previously using.
Some of the arguments have been removed.  So we have to update how we're
calling pip.

Next up we have the fact that dateutils uses pep517 as part of its build
process, which is fairly new.  This introduces a few wrinkles for us.
First, it utilizes "isolated builds", which should contain "only the
standard library and any explicitly requested build-dependencies."
Normally this is a good thing, but it's problematic for us because this
is an offline installer and we don't want pip creating a new build
environment for us and downloading dependencies from pypi.  So we have to
disable that feature (that's the `--no-build-isolation`).  Fortunately,
all of dateutils build deps we already have installed in our main
installation venv we create so we're all all set there.

As part of dateutils build process, pip shells out to other processes to
invoke various hooks.  It uses the same virtualenv by using
`sys.executable` when it invokes its python hooks.  Except that there
appears to be a bug, only for py37 as far as I can tell, where
`sys.executable` is wrong and instead of pointing to the virtualenv
python, it instead points to the system python.  As you might expect,
this causes the build/install to fail.  I've filed an issue on the
virtualenv repo to see what they think about it.

Turns out, using the built in `venv` module works as expected, in that
`sys.executable` is correctly set.  But of course `venv` is only
available in py33+.  Which brings us to our next change, conditionally
use `python -m venv` on py3, otherwise fall back to one of the versions
of external `virtualenv` on py26/py27.

This brings us to our final change, `python -m venv` does not install
wheel whereas `virtualenv` does install wheel.  So we need to install
wheel, conditionally, based on python version.

Going forward, there's a few things we should consider:

* py26 is being dropped real soon so a good chunk of this code can be
  removed.
* Once we can ensure we're using a modern version of pip after dropping
  py26, we should try switching to wheels instead of sdists where
  possible.  If a package has a py2/py3 compatible pure python wheel,
  (python-dateutil) we should use that instead of the sdist.  That
  way we can avoid all these build issues and just use the pre-built
  wheel.