Lock to dateutils v2.8.0 for bundled installer
For now we'll have to lock to 2.8.0 of dateutils due to python3.3
support.  Essentially, it doesn't seem like we can install dateutils
from source in python3.3 due to the fact that the version of
virtualenv/pip we'd need to leverage the build features dateutils
uses doesn't support python3.3.

This isn't normally an issue for people installing dateutils
through normal (online) pip because they can just use the prebuilt .whl
file which skips all the build steps.

I think once we drop py2.6/py3.3 support we can come back to this
again and:

1. Drop the constraints file.
2. Switch to pre-built wheels where possible (py2/py3 pure python).
   This should greatly simplify the install process.