Switch from safe to rt yaml loader
This allows us to preserve mapping order (which we were
already handling) as well as comments.

rt is a subclass of safe, from the docs:

https://yaml.readthedocs.io/en/latest/api.html

> The default loader (typ='rt') is a direct derivative of the safe loader,
without the methods to construct arbitrary Python objects that make the
unsafe loader unsafe, but with the changes needed for round-trip
preservation of comments, etc.. For trusted Python classes a constructor
can of course be added to the round-trip or safe-loader, but this has to
be done explicitly (add_constructor).

Closes #2757