Add shorthand parser error for duplicate keys set on same object
Currently it is not an error in the shorthand parser if you specify a
second key/value pair with the same key as a previous one. Only one of
these values will actually be set, depending on the version of python
it could be either the first or second.

This is a disproportionately larger issue than it seems (not a big issue
for JSON) because of unfamiliarity with the shorthand parser
syntax. There are also commonly used ec2 APIs that accept a list of
objects in the form Key=key,Value=value Key=key2,Value=value2. This is
a list of two dictionaries each with two keys Key and Value.

A common issue people run into is to write the previous example as
Key=key,Value=value,Key=key2,Value=value2 separating the list items
with a comma as well as the dictionary entries. This is now one
dictionary with two keys named Key and two keys named
Value. Previously this would be accepted and silently drop the
duplicates. Instead now you will get the following error:

Error parsing parameter '--attributes': Second instance of key "Key" encountered for input:
Key=key1,Value=value1,Key=key2,Value=value2
                      ^
This is often because there is a preceding "," instead of a space.