Always quote YAML boolean strings
CloudFormation treats unquoted values that are YAML 1.1 native booleans as
booleans, rather than strings. In YAML 1.2, the only boolean values are
"true" and "false" so values such as "yes" and "no" when loaded as strings are not
quoted when dumped. This logic ensures that we dump these values with quotes
so that CloudFormation treats these values as strings and not booleans.