Interpret unquoted booleans as Python booleans
This change ensures that if an unquoted boolean is provided to the
package command, the CLI will always output a boolean when it dumps it.
Without this change, the CLI will interpret the boolean as a string
and dump the value with quotes. We do not want this behavior because
CloudFormation respects some of the YAML 1.1 boolean values so
automatically converting them to a string may result in the incorrect
types being provided to the underlying AWS API call in the case
where a user is using a YAML boolean value other than true and false.

The one downside is that we will no longer roundtrip the specific
boolean value. For example, if "on" was specified, the package command
will dump it as "true". To get the desired behavior, we need to use
the RoundTripper classes, which currently is not completely
compatible with CloudFormation right now.