Bump python-rsa dependency range
PythonRSA 4.3 and 4.5 are the same release, they are also the same as
4.0 with some bugfix backports. Python 2 support was removed in 4.1,
but since 4.3/4.5 are just a re-released 4.0, they support python 2.7.

So the version range <=4.5 supports python 2.7 and python 3. However;
Python 3.4 support was also removed in 4.1, and is explicitly called
out as not restored in 4.3/4.5. Which means we need to branch our
dependency on rsa for Python 3.4 to include 4.0, which is the last
version to have Python 3.4 support. rsa 4.0 is also explicitlly pulled
into the bundled installer.