Add support for new parameter 'resource-type' for 'dlm create-default-role' command
'dlm create-default-role' command will be used to create the default role for AMI DLM policy.
Since the command is already used to create a default role for volume based DLM policy,
a new 'resource-type' parameter is being introduced to differentiate between which role
to create. This parameter is optional (for backward compatibility purposes) and defaults to
creating default role for volume based DLM policy.