Allow to name the session name when importing STSClientFactory to external code.
As the AWS EKS token mechanism is not implemented in boto3 / botocore directly (as it's actually a presigned URL workflow) some people want to import STSClientFactory as part of their own code.

```from awscli.customizations.eks.get_token import TokenGenerator, STSClientFactory```

For auditing purposes it's nice to have the possibility to submit a RoleSessionName other than EKSTokenAuth.

I'm not sure if this functionality is relevant for pure CLI use cases so I did not implement it.