Replace tests that rely on unreadable files with mocked timestamp
The timestamp tests do not work on all platforms, but the replacement
tests that rely on an unreadable file also do not work in some
environments. This commit reverts back to a timestamp based test, but
rather than relying on the acutal file timestamp we just mock out the
helper method that fetches it. We should replace these in the future
with something more robust that does not rely on a mock.