Remove "user level" NuGet.Config modifier from source updated message.
When adding new sources we can specify that we added the source to the user level NuGet.Config file. However, when updating an existing source we cannot be specific about which level NuGet.Config file was updated because it is possible that the existing source was not in the user level NuGet.Config. The 'nuget sources list' command returns all configured sources from all NuGet.Config levels. The 'nuget sources update' command updates the source in whichever NuGet.Config file the source was found.