Add WIP reimplementation of new-role wizard
This wizard provides a sample declaration of the new-role
wizard using the new UI.

It also required some updates to the wizard specification.
Specifically:
* Removes the notion of next steps for sections
* Add additional metadata to sections
* Allow prompts to have conditionals. This essentially replaces
  the next_steps notion.

Also note that because it starts with underscore, the wizard is
not propagated as a wizard command. In order to try out this wizard,
use the cli-dev wizard-dev command and specify it directly.