Extend select menu to be collapsable
This helpful when the select menu is being used as an input
container mixed with free form text. Specifically it will
expand the select menu when it is being focused but will collapse
it once the select menu loses focus, leaving its selection for
display.

In addition, the following changes were made to the original
SelectMenu:
* Allow items to be a callable. This allows for lazy loading of
  the items list.
* Do not manufacture the ui content width to be the same as the
  preferred width. This is important because the select menu will
  not always be its preferred width and can result in scroll bars
  being disconnected from the choices.