Introduce pytest into integration test suite
To start, this commit ports over a single s3 test class to use
pytest fixtures and style of assertions. A follow up commit
will be made to convert the rest of the tests to use pytest

It also introduced a new S3Utils abstraction that was pulled from
the logic in the BaseS3Transfer class. This was to be able to
move the s3 test suite off of the unittest classes and start
removing logic from testutils so that we are not including testing
logic in the awscli package proper as it is not included for the
CLI runtime.