Update make-exe to allow installation of wheels
This enables the building of the exe from a source directory that
has wheels in it. This in turn allows for the removal of
cryptography from the requirements-build.txt so it does not need
to be installed (via its wheel) prior to running the make-exe
script because the wheel can now be included in the source dir.