Add top-level import alias for botocore/s3transfer
This provides the following benefits in adding this alias:

* It minimizes churn on the entire code base as we do not have
  to update every import statement with regards to botocore and
  s3transfer
* It is easier to move implementations/patches between the CLI
  maintained implementation and the implementation for public
  consumption.
* It prevents breakages of any plugins that may be using botocore
  or s3transfer

The one downside is if anyone depends on the botocore or s3transfer
packages directly and imports from the awscli, it will result in
one of the implementations clobbering the other and likely unexpected
behavior. However, we have always discouraged users from importing
directly from the awscli package as it is not a support interface
of the AWS CLI.

It also required updating one of the tests to stop unnecessary
reloading of the awscli module. This was causing issues where the
CLI repository code was being referenced as the __file__ instead
of the install located in the site-packages.