Fix the help string for specifying multiple To, CC, BCC addresses
The To, CC, and BCC fields do support multiple addresses, but not as space separated values. The API would respond with an error: "An error occurred (InvalidParameterValue) when calling the SendEmail operation: Domain contains control or whitespace".

Instead, multiple address have to be specified with multiple --to fields, as alluded to in the Synopsis section further up the page.