Fix regression in not respecting --source-region
The CLI automatically does region redirects so while the
source region was not being set correctly, the CLI would detect
that the region is wrong on the client and redirect it to the
correct region. However, for some opt-in regions such as af-south-1,
region redirection is not supported which results in a hard fail.