Display error message when unable to open pager
When logic was added to lazily open the pager based on if there was
output from a CLI command, it introduced new behavior where no
output would be displayed if the pager could not be opened in
a subprocess. This is because the popen call on the pager was
now happening under the same try/except blocks that suppressed
IOError's related to the output stream closing prematurely, which
is also a parent class for the FileNotFoundError that is thrown
when the pager cannot be opened.

To fix this issue, the FileNotFoundError is converted to a new
custom error that allows us to control both the error message and
return code. However, it still does not propagate the command output
that would have been sent to the pager. In the future, we should consider
adding extra logic that prints both a warning message that the pager cannot
be opened and the output that would have been sent to the pager.
However, that would likely require a larger refactor.