Handle SIGPIPES properly
This commit changes the CLIDriver to consume potential SIGPIPE signals

In essence it will consume the signal and return a successful run.

This does changes process return code behaviour, as it will return a
zero code on broken pipes instead of previously returning a 120 code.
Because most of the pipelines only care about the return code of the
last process, and we can only receive a sigpipe if a next process in the
pipeline closes it, the return code from the aws cli in this case is not
really expressive

Before:

    % aws s3 ls | head -n 3 | cut -d ' ' -f 1
    2013-05-28
    2014-03-11
    2013-04-26

    [Errno 32] Broken pipe
    Exception ignored in: <_io.TextIOWrapper name='<stdout>' mode='w' encoding='utf-8'>
    BrokenPipeError: [Errno 32] Broken pipe

After:

    python3 bin/aws s3 ls | head -n 3 | cut -d ' ' -f 1
    2013-05-28
    2014-03-11
    2013-04-26

(the `cut` in the pipeline is just to remove bucket names)