Rely on build project to build wheels in CI
This is the currently recommended tool for following PEP517 to
build sdists and wheels. The command `python setup.py bdist_wheel`
is setuptools specific and does not take into account if we swap
out the build backend in the pyproject.toml.

Note we are building both the sdist and wheel when using the
`build` CLI because that is the recommendation so that you can
make sure that a wheel can be built from an sdist.