Update coverage configuration
The report was collecting coverage on both the installed
distribution and the git repository (which was not installed and
resulted in a coverage of zero). This change enforces only
coverage of the installed distribution is included in the report
so that downstream coverage services (e.g., codecov) does not
report the project as having zero coverage because it was looking
at the git repository coverage instead of the installed
distribution coverage.