Add in-tree PEP 517 build backend
This allows the auto-complete index to be automatically built as
part of building the wheel so that you no longer have to run the
gen-ac-index script in order to get a completely functional AWS
CLI when installing from source. Note that for editable installs,
it is still required to run the gen-ac-index script because
PEP 517 does not account for editable installs. However, once more
tools start supporting PEP 660 (which amends the PEP 517 interface
to support editable installs), we can drop this requirement.

Another side-effect to note is that because tools like pip will
try to build the CLI in an isolated virtualenv, installs will likely
fail if the tool's respective turn off build isolation flag is not
set (e.g., ``--no-build-isolation`` for pip). This is because the tool
will try to install the AWS CLI v2's implementation of botocore but
cannot because it is currently not on PyPI. I updated the developer
instructions in the README to include the ``--no-build-isolation``
flag and also updated the requirements.txt to include all runtime
dependencies, but it may trip people up until we have a better
install story around the AWS CLI v2's botocore implementation.