Use built-in ac index in completer functional test
Before, the auto-complete index would be generated and injected
into the completer during the functional tests. However, since
the index is being built as part of building the wheel, we can
now rely on the actual built index when we do our functional
smoke tests for the completer. This provides additional checks
that the completer is correctly using the index provided via
the package distribution.