Add version ceilings for build dependencies
The ceiling ensures that future versions of both setuptools and
wheel do not break building the distributions as the in-tree
build backend leverages both of these libraries. For setuptools,
it follows semver so its version ceiling is pinned to its
latest major version. For wheel, it does not have any a stable
API so its version ceiling is pinned to its latest release.