Add tests for in-tree backend
The purpose of the tests are to make sure we adhere to PEP 517 and
make sure the generated distributions look correct. A separate
run-backend-tests script was added for a couple of reasons:

1. You can separately test to make sure the backend works without
the awscli package installed as a distribution.

2. We need to turn off removing the repo root from the path because
the backends package is not included in the wheel that is installed
in site-packages and we have to import from the repository directly.