Remove aws_legacy_completer
The aws_legacy_completer logic was v1's implementation of the
aws_completer. This is was completely rewritten for v2 and the old
v1 logic is no longer supported. The script and logic only stuck around
in order to compare the v1 logic with the v2 logic during the development
of v2. We just never followed up to actually remove the script altogether
now that the new implementation was launched as the de facto completer
for v2.

We are removing it now so that individuals installing or packaging the
AWS CLI v2 from source do not have to deal with the script as it has
never been a supported interface of v2.