Add xdist pytest plugin
Spawns workers in subprocesses to run tests in order to speed up running
all of the tests at once. Some tests needed to be updated so that the
parameterized cases have a stable order as the workers will collect
all of the cases upfront and make sure the collected tests and order
is the same tests before running.

With regards to configuration:

* The number of processes is set to auto so it spawns workers based on
  the number of CPU cores. We cap this to four processes to make sure
  too much memory is not consumed in the case the environment has a lot
  of cores.

* The tests are configured to be distributed by test file to minimize the
  number of duplicate setups/teardowns of tests across the workers.