Port logic to prompt-toolkit 3
The main difference is that PTK3 uses the built-in Python
asyncio event loop. This required a larger refactor of the
test scafolding as some of the synchronous assumptions were no
longer valid. With regards to refactoring, I pivoted it to spinning
up the application in a thread and then feed input keys to the
app and make assertions on it while the app was running. This was
also cleaner in that we did not have to add actions and assertions
to a queue to be processed and instead can just make the assertion/action
directly.