Add additional check for preexisting completions
Having a complete_state does not necessarily mean that the
completions have been generated. The ptk completion process
starts the async completion generation with a blank
CompleteState object that has no completions. So in the check
before waiting, we should make sure the completions list is
populated if we are going to return before trying to wait.