Wait an invalidation cycle after app is running
Some of the tests assert the app state right after it is running.
However for the UI changes to propagate, especially around
conditional containers, it requires at least an additional rendering
cycle. So by manually calling invalidate() before yielding to make
assertions we can be more confident the changes have been propagated
to the UI.