autogenerate include-patterns for parent directories
Any include-path deep in a directory structure implicitly includes its
parent directories (though not necessarily the other contents of those
parents).

Currently, this does not matter, since the FileGenerator examines the
entire tree and emits all deeply-included paths. In a subsequent patch,
however, we want to restrict the FileGenerator from traversing excluded
paths, so we need to explicitly include the parents of any
deeply-included paths, or else a more shallow exclude will inhibit
discovery of the deeply-included paths, even if that exclude is
specified earlier in the command line.