Prefer man directly over groff
`groff` is great, as it easily generates ASCII text for help pages, but
lacks the escape sequences of `man` pages (bold, italic, etc.). Since
it's evident that `aws` help pages are formatted in the same way as
manual pages, I would expect it to look and feel the same way. For this
reason, instead of piping output through `groff` and then the user's
pager of choice, we now favor `man` (when available) to format these
pages in the standard way (which automatically uses the correct pager).