Port botocore compat module to v2
Same as the httpsession module, this module should be more or
less the same as what we use for v1. This was needed in order to
use the changes from the updated httpsession module. The only
difference is that this commit removes logic around CRT and
gzip models as we do not require it in v2.