Include global options in command docs
We get a lot of requests from users to include
global options in a command's help page. This PR
fulfils them by pre-generating .rst files for
global options and synopsis and writing them to
commands' help docs.

An alternative that was considered attempted to
plumb the global arg table from the CLIDriver to
any help command that would need it. However, we
abandoned the approach because it was too invasive
to the existing interfaces and there was no easy
way to apply the changes to customizations.